refactor: rename file decisionDefinition to decisionDefinitions
